Move scene, elements, and history into their respective modules
- Add "elements" argument to all scene functions to break reliance on global vars
- Move all scene functions under scene directory encapsulated by `index.ts`
- Move all element functions under `element` directory encapsulated by index.ts
- Create a SceneHistory class to store scene history state
- Add `createScene` function to create scene object
