+++ src/element/index.ts
@@ -1,0 +1,15 @@
+export { newElement } from "./newElement";
+export {
+  getElementAbsoluteX1,
+  getElementAbsoluteX2,
+  getElementAbsoluteY1,
+  getElementAbsoluteY2,
+  getDiamondPoints,
+  getArrowPoints
+} from "./bounds";
+
+export { handlerRectangles } from "./handlerRectangles";
+export { hitTest } from "./collision";
+export { resizeTest } from "./resizeTest";
+export { generateDraw } from "./generateDraw";
+export { isTextElement } from "./typeChecks";
