+++ src/element/typeChecks.ts
@@ -1,0 +1,7 @@
+import { ExcalidrawElement, ExcalidrawTextElement } from "./types";
+
+export function isTextElement(
+  element: ExcalidrawElement
+): element is ExcalidrawTextElement {
+  return element.type === "text";
+}
