+++ src/types.ts
@@ -1,0 +1,14 @@
+import { ExcalidrawElement } from "./element/types";
+
+export type AppState = {
+  draggingElement: ExcalidrawElement | null;
+  resizingElement: ExcalidrawElement | null;
+  elementType: string;
+  exportBackground: boolean;
+  currentItemStrokeColor: string;
+  currentItemBackgroundColor: string;
+  viewBackgroundColor: string;
+  scrollX: number;
+  scrollY: number;
+  name: string;
+};
